+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Added `Column.within_group()` for sorting the rows by the specified columns with some aggregation functions.
 - Added `Table` class, with the `update` and `delete` method to update and delete rows of a table in Snowflake.
   Also, the `merge` method is added to merge a `Table` with a `DataFrame`.
+- Added `DataFrame.cache_result()` for caching the operations performed on a `DataFrame` in a temporary table.
+  Subsequent operations on the original `DataFrame` have no effect on the cached result `DataFrame`.
 
 ## 0.3.0 (2022-01-09)
 ### New Features